feat: establish groups to handle objects
Created two groups, `updatable` and `drawable`, to make handling objects
a little be easier.  Added the Player object to each one and discovered
that the containers are applied to a specific _Class_ of object(s), and
not individual instances.

In the future, keep in mind that establishing groups must be done in a
specific order:
1. Create the groups.
2. Set them as containers for the desired _Class(es)_.
3. Create the objects within the classes.